### What problem does this PR solve?

Issue Number:   close https://github.com/pingcap/tidb/issues/17849

### What is changed and how it works?

Proposal: [executor: update the usage of the channel to improve the produce-consume relationship among operators](https://github.com/pingcap/tidb/issues/17849) <!-- REMOVE this line if not applicable -->

What's Changed:

- change ``HashJoinExec.probeResultChs`` and ``joinChkResourceCh`` from multi channel to single channel
- fix scope ambiguity of ``joinResult`` in ``joinMatchedProbeSideRow2ChunkForOuterHashJoin`` and ``joinMatchedProbeSideRow2Chunk``


### Check List

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)

```
name                                                                                                         old time/op    new time/op    delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12     1.29s ± 1%     1.32s ± 3%  +2.28%  (p=0.000 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12       268ms ± 1%     277ms ± 0%  +3.29%  (p=0.000 n=8+9)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        6.41s ± 3%     6.24s ± 6%  -2.59%  (p=0.043 n=10+10)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                  63.5µs ± 5%    60.9µs ± 4%  -4.11%  (p=0.006 n=10+9)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12           36.6ms ± 3%    36.3ms ± 1%    ~     (p=0.165 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12             32.6ms ± 2%    32.3ms ± 1%    ~     (p=0.546 n=9+9)

name                                                                                                         old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12    2.46GB ± 0%    2.46GB ± 0%    ~     (p=0.684 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12       307MB ± 0%     307MB ± 0%    ~     (p=0.579 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        903MB ± 0%     903MB ± 0%    ~     (p=0.684 n=10+10)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                  80.1kB ± 0%    79.3kB ± 0%  -0.99%  (p=0.000 n=10+9)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12           19.6MB ± 0%    19.6MB ± 0%    ~     (p=0.912 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12             17.9MB ± 0%    17.9MB ± 0%    ~     (p=0.684 n=10+10)

name                                                                                                         old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12      813k ± 0%      813k ± 0%  -0.01%  (p=0.033 n=10+9)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12        413k ± 0%      413k ± 0%  -0.01%  (p=0.002 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        2.01M ± 0%     2.01M ± 0%    ~     (p=0.343 n=10+10)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                     908 ± 0%       886 ± 0%  -2.42%  (p=0.000 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12             614k ± 0%      613k ± 0%  -0.00%  (p=0.000 n=10+10)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12               413k ± 0%      413k ± 0%  -0.01%  (p=0.000 n=10+10)
```

Side effects

- Performance regression
    - Consumes less CPU
    - Consumes less MEM

### Release note 

- update the usage of the channel in hash join to improve the produce-consume relationship among operators
- fix scope ambiguity of ``joinResult`` in ``joinMatchedProbeSideRow2ChunkForOuterHashJoin`` and ``joinMatchedProbeSideRow2Chunk``