@jianzhiyao 2021-03-24
After testing, i find there is no need to change mode of channel from 1:n to 1:1.
Maybe there are no enhancement on changing mode of channel.

### Testing code
```
func worker(wg *sync.WaitGroup, done chan bool, ch <-chan bool) {
	for ok := true; ok; {
		select {
		case <-done:
			ok = false
		case <-ch:
			wg.Done()
		}
	}
	wg.Done()
}

func BenchmarkMultiChannel(b *testing.B){
	var wg sync.WaitGroup
	concurrency := 5
	done := make(chan bool)
	chs := make([]chan bool, concurrency)
	for i := 0; i < concurrency; i++ {
		chs[i] = make(chan bool, 1)
		go worker(&wg, done, chs[i])
	}

	wg.Add(b.N * concurrency)
	for i := 0; i < b.N; i++ {
		for j := 0; j < concurrency; j++ {
			go func(j int) {
				chs[j] <- true
			}(j)
		}
	}
	wg.Wait()

	wg.Add(concurrency)
	close(done)
	wg.Wait()

	for i:= range chs {
		close(chs[i])
		for range chs[i] {
		}
	}
}
```

### benchmark output
```
BenchmarkMultiChannel
BenchmarkMultiChannel-6   	  545498	      2279 ns/op
BenchmarkMultiChannel-6   	  521736	      2448 ns/op
BenchmarkMultiChannel-6   	  545468	      2427 ns/op
BenchmarkMultiChannel-6   	  522376	      2313 ns/op
BenchmarkMultiChannel-6   	  571411	      2548 ns/op
BenchmarkMultiChannel-6   	  571408	      2247 ns/op
BenchmarkMultiChannel-6   	  571414	      2350 ns/op
BenchmarkMultiChannel-6   	  571400	      2473 ns/op
BenchmarkMultiChannel-6   	  444430	      2457 ns/op
BenchmarkMultiChannel-6   	  599960	      2514 ns/op
-------
BenchmarkSingleChannel
BenchmarkSingleChannel-6   	  444444	      3085 ns/op
BenchmarkSingleChannel-6   	  399979	      3470 ns/op
BenchmarkSingleChannel-6   	  387091	      3485 ns/op
BenchmarkSingleChannel-6   	  363639	      3407 ns/op
BenchmarkSingleChannel-6   	  363640	      3413 ns/op
BenchmarkSingleChannel-6   	  374991	      3309 ns/op
BenchmarkSingleChannel-6   	  342853	      3459 ns/op
BenchmarkSingleChannel-6   	  375038	      3394 ns/op
BenchmarkSingleChannel-6   	  400004	      3507 ns/op
BenchmarkSingleChannel-6   	  352953	      3785 ns/op

```