@fzhedu 2021-03-24
It is nice to bench multiple channels vs. single channel. But we should clarify the work cases:
1. an input goroutine generates `w` tasks, and each task takes several secondes, like task_i{i%3 sec};
2. n workers fetch `w ` tasks and each sleeps {i%3 sec};
3. an output goroutine fetch work results from 2.

we can bench the case from following aspects:
1. w >> n , and n increase from 1 to 1000 or even higher
2. fluctuant input and output speed.
3. channel size, at lease n