@jianzhiyao 2021-04-06
> @jianzhiyao what's your conclusion after the new benchmark?

###  Lastest benchmark
Seens changing channel mode is no a good practice for hash join. 

So I fix scope ambiguity of joinResult in https://github.com/pingcap/tidb/pull/23876. And I think https://github.com/pingcap/tidb/issues/17849 can be closed After https://github.com/pingcap/tidb/pull/23876 being merged.

PTAL @ichn-hu
```
name                                                                                                         old time/op    new time/op    delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12     1.29s ± 2%     1.33s ± 7%  +2.97%  (p=0.000 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12       265ms ± 1%     289ms ± 3%  +9.19%  (p=0.000 n=19+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        6.34s ± 2%     6.51s ± 4%  +2.69%  (p=0.000 n=20+20)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                  62.7µs ± 5%    64.4µs ± 5%  +2.78%  (p=0.012 n=20+19)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12           38.9ms ± 2%    37.7ms ± 3%  -3.08%  (p=0.000 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12             34.4ms ± 1%    33.9ms ± 2%  -1.39%  (p=0.000 n=20+20)

name                                                                                                         old alloc/op   new alloc/op   delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12    2.46GB ± 0%    2.46GB ± 0%    ~     (p=0.322 n=20+19)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12       307MB ± 0%     307MB ± 0%  -0.00%  (p=0.048 n=20+18)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        903MB ± 0%     903MB ± 0%    ~     (p=0.698 n=20+20)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                  80.1kB ± 0%    79.3kB ± 0%  -0.98%  (p=0.000 n=20+19)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12           19.6MB ± 0%    19.6MB ± 0%    ~     (p=0.841 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12             17.9MB ± 0%    17.9MB ± 0%    ~     (p=0.149 n=20+18)

name                                                                                                         old allocs/op  new allocs/op  delta
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12      813k ± 0%      813k ± 0%  -0.01%  (p=0.007 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12        413k ± 0%      413k ± 0%  -0.00%  (p=0.018 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_var_string(5)],_concurency:4,_joinKeyIdx:_[0],_disk:true)-12        2.01M ± 0%     2.01M ± 0%  -0.00%  (p=0.015 n=20+20)
HashJoinExec/(rows:5,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12                     908 ± 0%       886 ± 0%  -2.42%  (p=0.000 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-12             614k ± 0%      613k ± 0%  -0.00%  (p=0.000 n=20+20)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-12               413k ± 0%      413k ± 0%  -0.00%  (p=0.000 n=20+19)
```